Converted x,y coordinates to float
Now a robot's coordinates can either be specified with `int` or `float`,
and the infrastructure will support it. We move this, as it makes the
robot's position much more accurate.

+++ src/main/java/org/geeklet/machines/support/IRobot.java
@@ -8,8 +8,8 @@ import org.geeklet.machines.support.sensors.ISensor;
  * field must adhere.
  */
 public interface IRobot {
-    int getX();
-    int getY();
+    float getX();
+    float getY();
     int getDirection();
 
     void step();
